user_service/requirements/base.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.8.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pytz==2024.2
sqlparse==0.5.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
black==24.8.0
drf-yasg==1.21.7
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
flower==2.0.1
redis==5.0.8
setuptools==75.1.0
isort==5.13.2
flake8==7.1.1
